+++ src/test/kotlin/no/nav/sykdig/digitalisering/DigitaliseringsoppgaveServiceTest.kt
@@ -82,6 +82,7 @@ class DigitaliseringsoppgaveServiceTest : FellesTestOppsett() {
         beskrivelse = "Dette var ikkje bra",
         oppgaveType = OppgaveType.BEH_SED,
         tildeltEnhetsnr = "0393",
+        aktivDato = OffsetDateTime.now(),
     )
 
     val excpetedAvvisingsgrunn = Avvisingsgrunn.MANGLENDE_DIAGNOSE
